update to gorm v2

+++ db/condition/pagination.go
@@ -1,15 +1,15 @@
 package condition
 
-import "github.com/jinzhu/gorm"
+import "gorm.io/gorm"
 
 const globalDefaultPerPage = 30
 
 // Pagination 用來表示分頁
 type Pagination struct {
-	Page       int64 `query:"page" json:"page" url:"page" description:"目前頁面"`
-	PerPage    int64 `query:"perPage" json:"perPage" url:"perPage" description:"每頁顯示多少筆"`
-	TotalCount int64 `json:"totalCount" url:"-" description:"總筆數"`
-	TotalPage  int64 `json:"totalPage" url:"-" description:"總頁數"`
+	Page       int `query:"page" json:"page" url:"page" description:"目前頁面"`
+	PerPage    int `query:"perPage" json:"perPage" url:"perPage" description:"每頁顯示多少筆"`
+	TotalCount int `json:"totalCount" url:"-" description:"總筆數"`
+	TotalPage  int `json:"totalPage" url:"-" description:"總頁數"`
 }
 
 // Where return gorm scope function
@@ -19,7 +19,7 @@ func (p *Pagination) Where(db *gorm.DB) *gorm.DB {
 }
 
 // SetTotalCountAndPage 用來計算總數和分頁
-func (p *Pagination) SetTotalCountAndPage(total int64) {
+func (p *Pagination) SetTotalCountAndPage(total int) {
 	p.CheckOrSetDefault()
 	p.TotalCount = total
 
@@ -32,8 +32,8 @@ func (p *Pagination) SetTotalCountAndPage(total int64) {
 }
 
 // CheckOrSetDefault 檢查PerPage值若未設置則設置預設值
-func (p *Pagination) CheckOrSetDefault(params ...int64) *Pagination {
-	var defaultPerPage int64
+func (p *Pagination) CheckOrSetDefault(params ...int) *Pagination {
+	var defaultPerPage int
 	if len(params) >= 1 {
 		defaultPerPage = params[0]
 	}
@@ -52,12 +52,12 @@ func (p *Pagination) CheckOrSetDefault(params ...int64) *Pagination {
 }
 
 // LimitAndOffset return limit and offset
-func (p *Pagination) LimitAndOffset() (uint64, uint64) {
-	return uint64(p.PerPage), uint64(p.Offset())
+func (p *Pagination) LimitAndOffset() (int, int) {
+	return p.PerPage, p.Offset()
 }
 
 // Offset 計算 offset 的值
-func (p *Pagination) Offset() int64 {
+func (p *Pagination) Offset() int {
 	if p.Page <= 0 {
 		return 0
 	}
